Add files via upload
Wrote a generic class called HighestLowest with a type parameter constrained to any type that implements Comparable. The constructor accepts an array of such objects. The class has methods that return the highest and lowest object in the array. Demonstrated the class in an application by creating two instances of the HighestLowest  class, one with an array of Employee objects and another one with an array of City objects.